Adding handling of reading the pattern xml with two values of the meanDistPhi, both in the XMLConfigReader::readLUTs and ProcessorBase<GoldenPatternType>::configure
In the pattern xml now there are meanDistPhi0 and meanDistPhi1
attributes for each RefLayer.
In the XMLConfigReader::readLUTs the LUT address width for the
meanDistPhi is increased from 14 bits to 15. The
meanDistPhi[iLayer][iRefLayer][1] is stored in addresses >= 11520.
Backward compatible with the previous version of the meanDistPhi LUT and
pattern xml which contains only meanDistPhi