Working version with the threshold for the pdfSum value
GoldenPatternVec is changed to
std::vector<std::shared_ptr<GoldenPatternType> > 

getTreshold is added to the GoldenPattern. It is used in the
OMTFSorterWithThreshold.

GoldenPatternResult::finalise0() has now two version - for the classic
algorithm and the one with the thresholds. 



GhostBusterPreferRefDt - new ghost buster algorithm (in principle can be
used with any version of the algorithm)