Changes to allow using proper probabilities in the GoldenPattrens
GoldenPatternResult results in the GoldenPatternBase are now separate
for each processor (results is a 2d table).
GoldenPatternResult::finalise2() calculates the product of the
pdfVAlues.
OMTFSorterWithThreshold calculates the proper probability P(ptMu = gpPt
| delta_phis), which is stored in result.setGpProbability1(). For the
moment the applying of the threshold is commented, so the pattern with
the highest probability is selected.
PatternOptimizer allows to generated the pdf with the probabilities (and
not with the logarithms of probabilities): methods
updateStatCollectProb() and calulateProb(). It also allows to collect
the statistics of the GoldenPattern probabilities P(ptMu = gpPt |
delta_phis): method updateStatForAllGps (gpProbabilityStat in
GoldenPatternWithStat).
ProcessorBase<GoldenPatternType>::initPatternPtRange(bool firstPatFrom0)
allows now to set the range of the first two patterns (for the two
signs) from 0 Gev and not from the key().thePt (if firstPatFrom0 = true)