Added GhostBusterPreferRefDt, it provides better rate reduction
Added second implementation of the final ghostBuster-sorter:
GhostBusterPreferRefDt. If the count of the fired layers is the same,
the output candidates are sorted according to the number of the
reference layer. (In the original version (i.e. GhostBuster) the second
sorting condition was the sum of the pdfs). In this way, in the
GhostBusterPreferRefDt the candidate based on the DT reference layer has
priority over the one with the CSC or RPC reference layer. This provides
better rate reduction without drop of efficiency. To switch on the
GhostBusterPreferRefDt one have to add ghostBusterType =
cms.string("GhostBusterPreferRefDt") into simOmtfDigis_cfi.py