First version of the OMTFProcessorTTMerger - merging the TTTracks with muon chamber hits/segments.
OMTFProcessorTTMerger uses standard GoldenPattern to merge the TTTraks
with muon chamber hits/segments. When the reference layer is processed
(itGP->process1Layer1RefLayer ), then the phi of the reference hit is
replaced by the phi of the TTTrack at vertex, the GoldenPattern should
have appropriate pdf distribution for it.

OMTFProcessorTTMerger::ghostBust1 is not implemented yet
OMTFProcessorTTMerger::checkHitPatternValidity and getFinalcandidates
should be optimized for the TTTracks merging.

TTAlgoMuon has TrackingTriggerTrack - a wrapper for the TTTrack or
SimTrack (constructor from the TTTrack should be still implemented)

PatternGeneratorTT should be used to generate the patterns for the
OMTFProcessorTTMerger, for the moment it produces histograms of the
delta phi between the TTTrack at vertex and the reference hit vs pt of
the TTTrack.

Many additional changes:
- AlgoMuon keeps most of the values in GoldenPatternResult and
GoldenPatern* (and not in dedicated fields as before)
- moving calling the processor emulation steps from
OMTFReconstruction::getProcessorCandidates to
OMTFProcessor<GoldenPatternType>::run
- using vector of std::shared_ptr<AlgoMuon> in OMTFSorter and
GhostBuster