A few fixes in the algorithm improving efficiency and fake rate.
OMTFProcessorTTMerger:
GoldenPatternResult::finalise1() used in the OMTFProcessorTTMerger. This
functions simply counts the number of fired layers, without requiring
that in case of the DT both phi and phi_b is fired (i.e. pdf != 0).
ghostBust2 function used.
The candidates with quality 0 (i.e. with |eta| = 1.3 so outside of the
OMTF) are not return in the output collection.

GoldenPatternBase::process1Layer1RefLayer: now pdfVal = 0 returned when hit does not match