new feature: HSCP triggering
MuTimingModule allows to measure the track beta based on the muon stubs
timing (currently using only RPC).
MuTimingModuleWithStat is used to obtained needed timing distributions.
To the TrackingTriggerTrack and AlgoMuonBase simBeta is added, which is
probably temporary solution until something better is done.
MuonStub has roll field added, which is needed for obtaining beta from
timing.

In PdfModuleWithStats the sigma is increased by selected factor, which
is needed when the pdfs are obtained from the simTrack and not the
ttTracks. pdfModuleSimTracks100FilesSigma1p3.xml is currently the best
configuration.

In AlgoMuon getEta() renamed to getEtaHw()