Modification of the RpcClusterization to allow for dropping all clusters
from a given roll if there is more then rpcMaxClusterCnt clusters.

The dropping is done when rpcDropAllClustersIfMoreThanMax = true;

N.B. This modification is not fully correct, because the clusterization
(and thus dropping) is done for a RPC roll, and in the endcap one Link
Board is full chamber i.e. 3 rolls. So it is possible that e.g. there
are two clusters in one roll and one more in the other roll of the same
endcap chamber - then the cluster will not be dropped. To do it
correctly it is neccessary to introduce a structure corresponding to a
LB i.e. roll in case od the barrel and chamber in case of the endcap.

Above changes copied and integrated from the
from-CMSSW_10_6_0_pre4-phase2-KB_v1 commit
83d0ffaccb162f01d99c8d6ce4bc489673901256

Beside:
Prameterer lctCentralBx can be set for the process.simOmtfDigis