new class EventCapture
EventCapture implements IOMTFEmulationObserver. Captures the simMuons
from the event, and the OMTFinput, algoCandidates and gbCandidates for
each processor.

To use it, set process.simBayesOmtfDigis.eventCaptureDebug =
cms.bool(True)

Some arguments of the methods in the IOMTFEmulationObserver changed,
thus changes in every class that implements IOMTFEmulationObserver.

In the OMTFProcessor::run the OMTFinput is created as shared_ptr to
allow efficient passing to the
IOMTFEmulationObserver::observeProcesorEmulation

GoldenPatternResult::inalizeFunction can be set now from the python with paramter goldenPatternResultFinalizeFunction