Used for generation the patterns Patterns_0x0009_oldSample_3_10Files.xml - the patterns that contain the likelihood of hit missing in each layer.
Fixes in the XMLConfigReader - previously when the patterns were read
from multiple xml files, the patter groups where not set properly, what
affected patterns generation.

XMLConfigReader can read now the Patterns_template.xml that can contain
only the GP tags.

Rejecting of the phiB for the low quality DT is now applied already in
the GoldenPatternBase::process1Layer1RefLayer

GoldenPatternResult::finalise7() is devoted for using with the new patterns, it adds the pdfVal to the pdfSum regardless of the stub validity, it does not reject the loq quality DT phiB (as the finalise6), because the GoldenPatternBase::process1Layer1RefLayer must do it.