adding ghost buster byLLH and byFPLLH (Fired Planed - LogLikeliHood) and fixing quality table switching
in GhostBusterPreferRefDt::select added customLessByLLH and
customLessByFPLLH
Fixes in the OMTFConfiguration::configureFromEdmParameterSet and
OMTFProcessor<GoldenPatternType>::init

OMTFProcessor<GoldenPatternType>::getFinalcandidates  now it is
possible to swtich between the the Run2 quality tables (fwVersion 4, 6,
and 6) and the one proposed for the Run3 (fwVersion 8). Remember to
select the appropriate hwToLogicLayer_0x000*.xml in the python config.