EventCapture - stubs matching to the *SimLinks
It turned out that the SimHits are not stored for the pileup events (the
same as the SimTracks), so are useless for the stubs - simTracks
matching. So the rpcDigiSimLinks, cscStripDigiSimLinks and
dtDigiSimLinks are used, together with the trackingParticles. This
matching in the EventCapture is not finished yet, but it looks that the
*SimLinks are working.

CandidateSimMuonMatcher uses trackingParticles now

everywhere simTrackInputTag is used (instead e.g. g4SimTrackSrc)