Fix in CandidateSimMuonMatcher for the discplaced muons matching
CandidateSimMuonMatcher.cc
simTrackFilter uses now simTrackIsMuonInBx0, which do not apply cut on
the simTrack eta. Instead in match the eta of the propagated track to
the stattion2 is checked, whether it is inside the OMTF range. This
should work also with the displaced muons.

In many places abs changed to fabs where needed (even though the abs was
working here ok)