Merging l1t-phase2-v3.4.44 with from-phase2_CMSSW_11_1_7_KB_p2GmtV1
The base of this version is cms-l1t-offline:l1t-phase2-v3.4.44 - it
contains new lt_tpsID (Constants.h) luts tunned by Maria Cepeda (in
order to decrease the rate of the low pt muons) and a few other smaller
changes (?), but the barrelLimit0_ (0...4) are were not fixed there.
I copied into this versions the changes I made erlier in my branch
phase2_CMSSW_11_1_7_KB_p2GmtV1, e.g. the improved ghost busting
(cleanMuonV1) and DataDumper. The the barrelLimitX_ are fixed - which is
not necessary good because the lt_tpsID were tunned with the orginal
barrelLimitX_
In this commit the LogTrace were not yet copied from
phase2_CMSSW_11_1_7_KB_p2GmtV1