squash of 11 commits from from-CMSSW_13_1_0_KB_v1_displMu_LUTNN
Jan 8 - 29 Feb 2024

fixes in PatternGenerator and DataROOTDumper2

PatternGenerator.cc - different DistPhiBitShift for phase-1 and phase-2

DataROOTDumper2.cc - rejecting events with more than one matchingResults
(to clean the date for the NN training)
cleanStubs - fasle, but filtering the matchingResults

DataROOTDumper2.cc - dropping events with more than one matchingResults
small changes
Fixes in CandidateSimMuonMatcher

CandidateSimMuonMatcher.h
- added method matchSimple() that matches simTracks and OMTF cands
without propagation - it works well for the single muon samples. The
problem with propagation is that it does not work for low pt muons (pt <
~3GeV). Added flag usePropagation to switch between matchSimple() and
match() with propagation.
- added phi folding for delta phi in ghostBust() (but it looks like it
does not change the results)
- in match() (version for the SimTrack), the MatchingResult is added
also for the track for which the propagation failed.
The changes should be still added to the match() version for the
tracking particle.
bug fixes in StubResult and GoldenPatternResult, change finalise11()

Omtf/AlgoMuon.h
bug fix in AlgoMuon::getPatternNum() - checking if goldenPaternUnconstr,
otherwise segmentation fault can be generated

StubResult.h
added method setPdfVal(), which is used in the
GoldenPatternResult::finalise11()

GoldenPatternResult.cc
GoldenPatternResult::finalise11() - added penalty -63 for the phiB
layers, if there was a hit but it does not fit the pdf
operator<< - bug fix - checking if the pResult.omtfConfig is nullptr -
this case occurs for the empty GoldenPatternResult

L1Trigger/L1TMuonOverlapPhase1/test/expert/omtf/runMuonOverlap_PatsWithExtrapolation_efficiencyAnalyser.py
- improved .py for running the emulator + analyzer
Fixes in DataROOTDumper2 and CandidateSimMuonMatcher

DataROOTDumper2 added flag usePropagation which if false turns on the
filter on matchingResults.size() > 1

CandidateSimMuonMatcher::match() - threshold updated to be good for
displaced muons in H2ll sample. There should be an automatic switch for
this threshold

runMuonOverlapWithExtrapol_phase2_efficiencyAnalyser_rootDumper.py
added flags matchUsingPropagationInAnlyzer  and
matchUsingPropagationInDumper
fixes of the candidate hwPt and HwPtUnconstrained

Omtf/AlgoMuon.h - changed the value of getPtConstr and getPtUnconstr for
an empty candidate from -1 to 0 (as it is in the interface note)

OMTFConfiguration::etaBits2HwEta - added return 95 for bits ==
0b111111111, as it is like that tin the FW

OMTFProcessor<GoldenPatternType>::getFinalcandidates - hwPt is set to 1
if the constraint getPtConstr is 0 but getPtUnconstr > 0

L1Trigger/L1TMuonOverlapPhase1/src/Omtf/OMTFSorter.cc - bestGP is set to
the golden pattern number 0 if getPdfSum() == 0

New version of the pT regression neural network

The NN is trained on the top of the phase-2 OMTF pattern logic.

LutNeuronLayerFixedPoint:
added template parameter output_F - this allows to have better precision
of the lut out sum calculation, and thus better precision of the next
layer input.
runWithInterpolation - fixed protection against overflow of the LUT
address,
getOutWithOffset() -  fixed protection against overflow of the output
value.

LutNetworkFixedPointRegression2Outputs - added template parameters
layer1_output_F,  layer3_input_F.

pt calibration with linear scaling of the output pt instead of
ptCalibrationArray.

PtAssignmentNNRegression.cc - different bit precision for the NN than in
the previous version of the NN.

omtfHitWithQualToEventInput() - DT stubs quality is used for the NN
input.

L1TMuonOverlapPhase1/src/Omtf/OMTFProcessor.cc - setting the output
unconstrained pT to 1 when getPdfSumConstr.

applied scram build code-checks and scram build code-format

t22
small fixes

OMTFProcessor<GoldenPatternType>::getFinalcandidates - fixing the
condition for the validity of the constrained pt

LutNetworkFixedPointRegression2Outputs.h run() - constrainig noHitsCnt
to 15
PtAssignmentNNRegression::getPts - small fixes

XMLEventWriter.cc - removed constrain in the number of events that can
be dumped to the xml